simworld/backend/app/main.py
import os
from dotenv import load_dotenv
load_dotenv(os.path.join(os.path.dirname(__file__), "..", "..", ".env"), override=True)

from fastapi import FastAPI
from fastapi.middleware.cors import CORSMiddleware

from app.config import get_settings
from app.api import simulations, health, auth, billing, network

settings = get_settings()

app = FastAPI(
    title="SimWorld API",
    description="Scenario simulation and prediction platform powered by MiroFish",
    version="0.1.0",
)

app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=False,
    allow_methods=["*"],
    allow_headers=["*"],
)

app.include_router(health.router, tags=["health"])
app.include_router(simulations.router, prefix="/api/simulations", tags=["simulations"])
app.include_router(auth.router, prefix="/api/auth", tags=["auth"])
app.include_router(billing.router, prefix="/api/billing", tags=["billing"])
app.include_router(network.router, prefix="/api/simulation", tags=["network"])
