simworld/backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
celery[redis]==5.3.6
redis==5.0.1
httpx>=0.24,<0.26
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0
supabase==2.3.4
stripe==8.4.0
python-jose[cryptography]==3.3.0
fpdf2==2.8.3
svix>=1.8.1
anthropic>=0.39.0
